apps/tests/selenium/requirements.txt
# Selenium Robot Framework Test Requirements
selenium==4.15.2
robotframework==6.1.1
robotframework-seleniumlibrary==6.2.0
robotframework-requests==0.9.7
robotframework-faker==5.0.0
robotframework-databaselibrary==1.2.4
robotframework-pabot==2.16.0
requests==2.31.0
pytest==7.4.3
pytest-html==4.1.1
webdriver-manager==4.0.1
pillow==10.1.0
openpyxl==3.1.2
python-dotenv==1.0.0